feat(DPE-6716): LDAP Events (#29)
## 🏷 Type of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Documentation Update
- [ ] Tooling and CI changes
- [x] Dependencies upgrade or change

## 📝 Description
LDAP events and refactor a bit the management of restarts
This PR introduces the `LDAPEventHandler` which handles the events of the three following relations:
 * ldap
 * ldap-certificate-transfer
 * ldap-peers

It also introduces a custom ops event that is used to provide a single source of truth for restarts in the LDAP integration.
There is a bit of rework of the restarts, and an improvement of the get_status that now also checks for the ldap connection by doing a bind operation.

## 📑 Motivation and Context
This is the implementation of [DPE-6716](https://warthogs.atlassian.net/browse/DPE-6716) for the ldap integration.
It adds the support for the events.
The next steps will be to update the config so that the services are configured to integrate with LDAP.

## 🧪 How Has This Been Tested?
This has been manually tested for now.
I'm currently adding unit tests.

## ✅ Checklist
- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the [**CONTRIBUTING**](../blob/6/edge/CONTRIBUTING.md) document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.


[DPE-6716]: https://warthogs.atlassian.net/browse/DPE-6716?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ