feat: LDAP configuration (#32)
This PR updates the configuration managers for mongodb and mongos to conditionally add the ldap configuration for the service.
It also adds some testing for this configuration and some full cycle unit testing.
Finally, it updates the restart mechanism to restart only if needed.


* feat(DPE-6792): Share config with mongos (#33)

Instead of requiring the operator to config the two services, the config-server shares the ldapUserToDNMapping with mongos through the cluster relation.
It also shares a hash of the ca chain + ldap addresses to ensure that both services are connected with the same ldap server.

---------

Co-authored-by: Mia Altieri <32723809+MiaAltieri@users.noreply.github.com>